pyproject.toml
[project]
name = "anki-tagger"
version = "0.1.0"
description = "Tag Anki decks with learning objectives using embedding models"
readme = "README.md"
requires-python = ">=3.9.2"
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pandas>=1.0.0",
    "numpy>=1.20.0",
    "tiktoken>=0.5.0",
    "tqdm>=4.60.0",
    "pdfplumber>=0.7.0",
    "openai>=0.27.0",
    "openrouter>=0.1.0",
    "ollama>=0.1.0",
    "anki>=25.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = [
    "config_loader.py",
    "config.yaml",
    "embed_anki_deck.py",
    "embedding_client.py",
    "make_learning_objectives.py",
    "run_pipeline.py",
    "select_cards.py",
    "tag_deck.py",
]
